dns: divide test file in three parts
