+++ .gitignore
@@ -71,13 +71,14 @@ test/echo_client_evbuf
 test/net/buffer
 test/net/connection
 test/net/connection_check_fds
-test/net/dns
 test/ooni/dns_injection
 test/ooni/net_test
 test/ooni/http_invalid_request_line
 test/ooni/tcp_connect
 test/ooni/tcp_test
-test/dns/dns
+test/dns/query
+test/dns/resolver
+test/dns/response
 test/http/http
 test/report/file
 test/traceroute/android
